Refactor utils and services
