+++ frontend/src/hooks/realtime/useRealtimeCalendars.js
@@ -1,8 +1,8 @@
 import { useEffect, useContext, useRef } from 'react';
-import { supabase } from '../../services/supabaseClient';
+import { supabase } from '../../services/supabase/supabaseClient';
 import { log } from '../../utils/logger';
 import { UserContext } from '../../contexts/UserContext';
-import { analyticsPromise } from '../../services/firebase';
+import { analyticsPromise } from '../../services/firebase/firebase';
 import { logEvent } from 'firebase/analytics';
 
 const API_URL = import.meta.env.VITE_API_URL;
