+++ frontend/src/hooks/realtime/useRealtimeTokens.js
@@ -1,9 +1,9 @@
 import { useContext, useEffect, useRef } from 'react';
 import { UserContext } from '../../contexts/UserContext';
-import { analyticsPromise } from '../../services/firebase';
+import { analyticsPromise } from '../../services/firebase/firebase';
 import { log } from '../../utils/logger';
 import { logEvent } from 'firebase/analytics';
-import { supabase } from '../../services/supabaseClient';
+import { supabase } from '../../services/supabase/supabaseClient';
 
 const API_URL = import.meta.env.VITE_API_URL;
