+++ frontend/src/pages/calendar/CalendarView.jsx
@@ -6,8 +6,8 @@ import dayGridPlugin from '@fullcalendar/daygrid';
 import interactionPlugin from '@fullcalendar/interaction';
 import { useTranslation } from 'react-i18next';
 import { UserContext } from '../../contexts/UserContext';
-import { formatToLocalISODate, getMondayFromDate } from '../../utils/dateUtils';
-import { getCalendarSourceMap } from '../../utils/calendarSourceMap';
+import { formatToLocalISODate, getMondayFromDate } from '../../utils/calendar/dateUtils';
+import { getCalendarSourceMap } from '../../utils/calendar/calendarSourceMap';
 import ShareCalendarModal from '../../components/calendar/ShareCalendarModal';
 import AlertSystem from '../../components/common/AlertSystem';
 import isEqual from 'lodash/isEqual';
