+++ frontend/src/utils/api/fetchSuggestions.js
